backend/requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
pydantic
sqlalchemy
alembic
httpx
slowapi
pytest
pytest-asyncio
ruff
black
